clauseX/requirements.txt
fastapi
uvicorn
faiss-cpu
uvicorn[standard]
pandas
mysql-connector-python
python-dotenv
pydantic
langchain
langchain-community
langchain-google-genai
langchain-groq
python-multipart
